rename and move the folder
